requirement.txt
torch==2.0.0
torchvision==0.15.0
transformers
mlflow
streamlit
Pillow
